Enhance UI Responsiveness and Navigation
