+++ eloh-app/app/ourTeam/page.jsx
@@ -1,13 +1,13 @@
 "use client";
 
-import Navbar from "@/components/maiNavbar";
+import Navbar from "@/components/MainNavbar";
 import React from "react";
 
 const OurTeam = () => {
     return (
         <>
-        <Navbar />
-        
+            <Navbar />
+
             <h1 className="flex justify-center mt-20 text-2xl font-bold">Our Team</h1>
         </>
     )
